+++ tests/e2e/resources/expected_downloads_summaries/plugins/test_audio_extract_single.json
@@ -1,3 +1,3 @@
 {
-  "YouTube Rewind 2019： For the Record ｜ #YouTubeRewind.mp3": "2b44b454ecc58f3600731a44e0c61f1b"
+  "YouTube Rewind 2019： For the Record ｜ #YouTubeRewind.mp3": "c4ce74e2ddc4e6a0ab823b5b622e2275"
 }
